requirements.txt
streamlit
pandas
bcrypt
plotly
requests
python-dotenv
